+++ src/dtos/route.rs
@@ -5,7 +5,7 @@ use validator::Validate;
 
 use crate::{
     models::route::{Route, RouteStatus},
-    utils::utils::is_valid_uuid,
+    utils::string::is_valid_uuid,
 };
 
 #[derive(Validate, Debug, Default, Clone, Serialize, Deserialize)]
