+++ src/dtos/vehicle.rs
@@ -3,7 +3,7 @@ use serde::{Deserialize, Serialize};
 use validator::Validate;
 
 use crate::models::vehicle::Vehicle;
-use crate::{models::vehicle::VehicleDocument, utils::utils::is_valid_uuid};
+use crate::{models::vehicle::VehicleDocument, utils::string::is_valid_uuid};
 
 #[derive(Validate, Debug, Default, Clone, Serialize, Deserialize)]
 #[serde(rename_all = "camelCase")]
